Rename Checkbox.jsx to CheckboxWidget.jsx to match the naming on the original
https://github.com/mozilla-services/react-jsonschema-form/blob/158fd6465b46ba232864727a05704bf3708d2c5e/src/components/widgets/CheckboxWidget.js
