+++ packages/api-explorer/src/form-components/CheckboxWidget.jsx
